# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LongInvest",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/quote-cycles": {
      "get": {
        "tags": [
          "quotes"
        ],
        "summary": "List Cycles",
        "operationId": "list_cycles_api_v1_quote_cycles_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/QuoteCycleStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteCyclePageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quote-cycles/{cycle_id}/items": {
      "get": {
        "tags": [
          "quotes"
        ],
        "summary": "List Cycle Items",
        "operationId": "list_cycle_items_api_v1_quote_cycles__cycle_id__items_get",
        "parameters": [
          {
            "name": "cycle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cycle Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 200,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteItemsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quote-cycles/manual": {
      "post": {
        "tags": [
          "quotes"
        ],
        "summary": "Submit Manual Cycle",
        "operationId": "submit_manual_cycle_api_v1_quote_cycles_manual_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 160,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualQuoteRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quotes/diagnose": {
      "post": {
        "tags": [
          "quotes"
        ],
        "summary": "Diagnose Quotes",
        "operationId": "diagnose_quotes_api_v1_quotes_diagnose_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 160,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnoseQuoteRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiagnoseQuoteRequest": {
        "properties": {
          "symbols": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 200,
            "minItems": 1,
            "title": "Symbols"
          },
          "confirm": {
            "type": "boolean",
            "title": "Confirm"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "symbols",
          "confirm"
        ],
        "title": "DiagnoseQuoteRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManualQuoteRequest": {
        "properties": {
          "symbols": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 200,
            "minItems": 1,
            "title": "Symbols"
          },
          "timeout_seconds": {
            "type": "integer",
            "maximum": 60.0,
            "minimum": 10.0,
            "title": "Timeout Seconds",
            "default": 30
          },
          "confirm": {
            "type": "boolean",
            "title": "Confirm"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "symbols",
          "confirm"
        ],
        "title": "ManualQuoteRequest"
      },
...
      "QuoteCyclePageData": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QuoteCycleRecord"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "QuoteCyclePageData"
      },
      "QuoteCyclePageResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "title": "Success"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "server_time": {
            "type": "string",
            "format": "date-time",
            "title": "Server Time"
          },
          "data": {
            "$ref": "#/components/schemas/QuoteCyclePageData"
          }
        },
        "type": "object",
        "required": [
          "success",
          "code",
          "message",
          "request_id",
          "server_time",
          "data"
        ],
        "title": "QuoteCyclePageResponse"
      },
      "QuoteCycleRecord": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/QuoteCycleStatus"
          },
          "expected_count": {
            "type": "integer",
            "title": "Expected Count"
          },
          "valid_count": {
            "type": "integer",
            "title": "Valid Count"
          },
          "missing_count": {
            "type": "integer",
            "title": "Missing Count"
          },
          "conflict_count": {
            "type": "integer",
            "title": "Conflict Count"
          },
          "failed_count": {
            "type": "integer",
            "title": "Failed Count"
          },
          "eligible_item_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Eligible Item Ids"
          },
          "eligible_symbols": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Eligible Symbols"
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "deadline_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline At"
          },
          "finalized_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finalized At"
          },
          "schedule_occurrence_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule Occurrence Id"
          },
          "subscription_snapshot_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subscription Snapshot Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "expected_count",
          "valid_count",
          "missing_count",
          "conflict_count",
          "failed_count",
          "eligible_item_ids",
          "eligible_symbols",
          "scheduled_at",
          "started_at",
          "deadline_at",
          "finalized_at",
          "schedule_occurrence_id",
          "subscription_snapshot_version"
        ],
        "title": "QuoteCycleRecord"
      },
      "QuoteCycleStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "FETCHING",
          "FINALIZING",
          "READY",
          "PARTIAL",
          "FAILED",
          "MISSED",
          "CANCELED"
        ],
        "title": "QuoteCycleStatus"
      },
...
      "QuoteItemRecord": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "cycle_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cycle Id"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "price": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "open": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open"
          },
          "high": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "High"
          },
          "low": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Low"
          },
          "previous_close": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Close"
          },
          "volume": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          },
          "amount": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "quote_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quote Time"
          },
          "received_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received At"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "conflict_evidence": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conflict Evidence"
          },
          "eligible_for_evaluation": {
            "type": "boolean",
            "title": "Eligible For Evaluation"
          },
          "expected_subscription_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Subscription Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "cycle_id",
          "symbol",
          "status",
          "price",
          "open",
          "high",
          "low",
          "previous_close",
          "volume",
          "amount",
          "quote_time",
          "received_at",
          "provider",
          "error_code",
          "conflict_evidence",
          "eligible_for_evaluation",
          "expected_subscription_version"
        ],
        "title": "QuoteItemRecord"
      },
      "QuoteItemsData": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QuoteItemRecord"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "QuoteItemsData"
      },
      "QuoteItemsResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "title": "Success"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "server_time": {
            "type": "string",
            "format": "date-time",
            "title": "Server Time"
          },
          "data": {
            "$ref": "#/components/schemas/QuoteItemsData"
          }
        },
        "type": "object",
        "required": [
          "success",
          "code",
          "message",
          "request_id",
          "server_time",
          "data"
        ],
        "title": "QuoteItemsResponse"
      },
      "QuoteJobData": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "QuoteJobData"
      },
      "QuoteJobResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "const": true,
            "title": "Success"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "server_time": {
            "type": "string",
            "format": "date-time",
            "title": "Server Time"
          },
          "data": {
            "$ref": "#/components/schemas/QuoteJobData"
          }
        },
        "type": "object",
        "required": [
          "success",
          "code",
          "message",
          "request_id",
          "server_time",
          "data"
        ],
        "title": "QuoteJobResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}